+++ src/hsp3/dpmread.h
@@ -5,14 +5,14 @@
 #ifndef __dpmread_h
 #define __dpmread_h
 
-int dpm_ini( char *dpmfile, long dpmofs, int chksum, int deckey, int slot=0 );
+int dpm_ini( char *dpmfile, size_t dpmofs, int chksum, int deckey, int slot=0 );
 void dpm_bye( void );
 
 FILE *dpm_open( char *fname );
-int dpm_fread( void *mem, int size, FILE *stream );
+size_t dpm_fread( void *mem, size_t size, FILE *stream );
 void dpm_close(FILE* fp);
-int dpm_read( char *fname, void *readmem, int rlen, int seekofs );
-int dpm_exist( char *fname );
+size_t dpm_read( char *fname, void *readmem, size_t rlen, size_t seekofs );
+size_t dpm_exist( char *fname );
 void dpm_getinf( char *inf );
 int dpm_filecopy( char *fname, char *sname );
 int dpm_filebase( char *fname );
