+++ src/hsp3/ios/supio_ios.h
@@ -3,7 +3,7 @@
 //	supio.cpp functions (linux)
 //
 
-char *mem_ini( int size );
+char *mem_ini( size_t size );
 void mem_bye( void *ptr );
 int mem_save( char *fname, void *mem, int msize, int seekofs );
 void strcase( char *str );
