+++ src/hsp3/win32gui/supio_win.cpp
@@ -25,7 +25,7 @@
 //
 //		basic C I/O support
 //
-char *mem_ini( int size ) {
+char *mem_ini(size_t size ) {
 	return (char *)calloc(size,1);
 }
 
@@ -33,7 +33,7 @@ void mem_bye( void *ptr ) {
 	free(ptr);
 }
 
-char *mem_alloc( void *base, int newsize, int oldsize )
+char *mem_alloc( void *base, size_t newsize, size_t oldsize )
 {
 	char *p;
 	if ( base == NULL ) {
