+++ src/hsp3/win32gui/supio_win.h
@@ -5,9 +5,9 @@
 
 #include "../hsp3utfcnv.h"
 
-char *mem_ini( int size );
+char *mem_ini(size_t size );
 void mem_bye( void *ptr );
-char *mem_alloc( void *base, int newsize, int oldsize );
+char *mem_alloc( void *base, size_t newsize, size_t oldsize );
 
 //void prtini( char *mes );
 //void prt( char *mes );
